+++ src/pyochain/abc/_iterator.py
@@ -3,14 +3,23 @@ from __future__ import annotations
 import functools
 import itertools
 from abc import ABC
-from typing import TYPE_CHECKING, Any, Literal, TypeGuard, TypeIs, overload
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Literal,
+    Self,
+    TypeGuard,
+    TypeIs,
+    overload,
+    override,
+)
 
 from .. import _tools as tls  # pyright: ignore[reportMissingModuleSource]
 from .._abc import (  # pyright: ignore[reportMissingModuleSource]
     PyoIterable,
     PyoIteratorRS,
 )
-from ..rs import Option, option
+from ..rs import NONE, Err, Null, Ok, Option, Result, Some, option
 
 if TYPE_CHECKING:
     from collections.abc import (
@@ -1250,40 +1259,46 @@ class PyoIterator[T](PyoIteratorRS[T], ABC):
             itertools.accumulate(iter(self), func, initial=initial)
         )
 
-    def peekable(self, n: int) -> tuple[Seq[T], PyoIterator[T]]:
-        """Retrieve the next **n** elements from the `Iterator`, whilst leaving the original iterator unconsumed.
+    def peekable(self) -> Peekable[T]:
+        """Creates an iterator which can use the peek and peek_mut methods to look at the next element of the `Iterator` without consuming it.
 
-        The returned tuple contains two elements:
+        See their documentation for more information.
 
-        - A `Seq` of the next **n** elements.
-        - An `Iterator` that includes the peeked elements followed by the remaining elements of the original `Iterator`.
+        Note that the underlying `Iterator` is still advanced when peek or peek_mut are called for the first time.
 
-        Args:
-            n (int): Number of items to peek.
+        In order to retrieve the next element, `next` is called on the underlying `Iterator`, hence any side effects (i.e. anything other than fetching the next value) of the `next` method will occur.
 
         Returns:
-            tuple[Seq[T], PyoIterator[T]]: A tuple containing the peeked elements and the remaining iterator.
+            Peekable[T]: A new `Iterator` that allows peeking at the next element.
 
-        See Also:
-            [`Iter::cloned`][cloned] to create an independent copy of the iterator.
-
-        Example:
+        Examples:
+            Basic usage:
             ```python
-            >>> from pyochain import Iter, Seq
-            >>> peeked, remaining = Iter((1, 2, 3)).peekable(2)
-            >>> peeked
-            Seq(1, 2)
-            >>> remaining.collect(Seq)
-            Seq(1, 2, 3)
+            >>> from pyochain import Range
+            >>> xs = Range(1, 4)
+            >>> iterator = xs.iter().peekable()
+            >>> # peek() lets us see into the future
+            >>> iterator.peek()
+            Some(1)
+            >>> iterator.next()
+            Some(1)
+            >>> iterator.next()
+            Some(2)
+            >>> # we can peek() multiple times, the iterator won't advance
+            >>> iterator.peek()
+            Some(3)
+            >>> iterator.peek()
+            Some(3)
+            >>> iterator.next()
+            Some(3)
+            >>> # after the iterator is finished, so is peek()
+            >>> iterator.peek()
+            NONE
+            >>> iterator.next()
+            NONE
 
-            ```
         """
-        from .._seq import Seq
-
-        iterator = iter(self)
-        peeked = Seq(itertools.islice(iterator, n))
-        remaining = self._from_iterable(itertools.chain(peeked, iterator))
-        return peeked, remaining
+        return Peekable(iter(self))
 
     def array_chunks(self, size: int) -> PyoIterator[PyoIterator[T]]:
         """Yield subiterators (chunks) that each yield a fixed number elements, determined by size.
@@ -2990,3 +3005,168 @@ class PyoIterator[T](PyoIteratorRS[T], ABC):
             ```
         """
         return self._from_iterable(tls.WithPosition(iter(self)))
+
+
+class Peekable[T](PyoIterator[T]):
+    __slots__ = ("_it", "_peeked")  # pyright: ignore[reportUnannotatedClassAttribute, reportIncompatibleUnannotatedOverride]
+
+    def __init__(self, iterable: Iterable[T]) -> None:
+        self._it: Iterator[T] = iter(iterable)
+        self._peeked: Option[T] = NONE
+
+    @override
+    def __iter__(self) -> Self:
+        return self
+
+    @override
+    def __next__(self) -> T:
+        match self._peeked:
+            case Some(value):
+                self._peeked = NONE
+                return value
+            case Null():
+                return next(self._it)
+
+    def __bool__(self) -> bool:
+        return self.peek().is_some()
+
+    def peek(self) -> Option[T]:
+        """Returns the `next()` value without advancing the `Iterator`.
+
+        Returns:
+            Option[T]: The next value wrapped in `Some(T)` if available, or `NONE` if the iteration is over.
+        """
+        match self._peeked:
+            case Some(_):
+                return self._peeked
+            case Null():
+                try:
+                    self._peeked = option(next(self._it))
+                except StopIteration:
+                    return NONE
+                else:
+                    return self._peeked
+
+    def next_if(self, func: Callable[[T], bool]) -> Option[T]:
+        """Consume and return the next value of this iterator if a condition is `True`.
+
+        Args:
+            func (Callable[[T], bool]): A function that takes the next value and returns a boolean.
+
+        Returns:
+            Option[T]: The next value wrapped in `Some(T)` if the condition is true, or `NONE` if the condition is false or the iteration is over.
+
+        Examples:
+            Consume a number if it's equal to 0.
+            ```python
+            >>> from pyochain import Range
+            >>> iterator = Range(0, 5).iter().peekable()
+            >>> # The first item of the iterator is 0; consume it.
+            >>> iterator.next_if(lambda x: x == 0)
+            Some(0)
+            >>> # The next item returned is now 1, so `next_if` will return `None`.
+            >>> iterator.next_if(lambda x: x == 0)
+            NONE
+            >>> # `next_if` retains the next item if the predicate evaluates to `false` for it.
+            >>> iterator.next()
+            Some(1)
+
+            ```
+            Consume any number less than 10.
+            ```python
+            >>> iterator = Range(1, 20).iter().peekable()
+            >>> # Consume all numbers less than 10
+            >>> while iterator.next_if(lambda x: x < 10).is_some():
+            ...     pass
+            >>> # The next value returned will be 10
+            >>> iterator.next()
+            Some(10)
+        """
+        match self.next():
+            case Some(matched) if func(matched):
+                return Some(matched)
+            case other:
+                self._peeked = other
+                return NONE
+
+    def next_if_eq(self, expected: T) -> Option[T]:
+        """Return the next item if it is equal to expected.
+
+        Returns:
+            Option[T]: The next value wrapped in `Some(T)` if it is equal to expected, or `NONE` if it is not equal or the iteration is over.
+
+        Example:
+            Consume a number if it's equal to 0.
+            ```python
+            >>> from pyochain import Range
+            >>> iterator = Range(0, 5).iter().peekable()
+            >>> # The first item of the iterator is 0; consume it.
+            >>> iterator.next_if_eq(0)
+            Some(0)
+            >>> # The next item returned is now 1, so `next_if_eq` will return `None`.
+            >>> iterator.next_if_eq(0)
+            NONE
+            >>> # `next_if_eq` retains the next item if it was not equal to `expected`.
+            >>> iterator.next()
+            Some(1)
+
+            ```
+        """
+        return self.next_if(lambda nxt: nxt == expected)
+
+    def next_if_map[R](
+        self,
+        f: Callable[[T], Result[R, T]],
+    ) -> Option[R]:
+        """Consumes the next value of this `Iterator` and applies a function *f* on it, returning the result if the closure returns `Ok`.
+
+        Otherwise if the closure returns `Err` the value is put back for the next iteration.
+
+        The content of the `Err` variant is typically the original value of the closure, but this is not required.
+
+        If a different value is returned, the next `peek()` or `next()` call will result in this new value.
+
+        Args:
+            f (Callable[[T], Result[R, T]]): A function that takes the next value and returns a Result.
+
+        Returns:
+            Option[R]: The result of the function wrapped in `Some(R)` if the function returns `Ok(R)`, or `NONE` if the function returns `Err(T)` or the iteration is over.
+
+        Examples:
+            Parse the leading decimal number from an iterator of characters.
+            ```python
+            >>> from pyochain import Iter, Option, Some, NONE
+            >>> import unicodedata
+            >>>
+            >>> iterator = Iter("125 GOTO 10").peekable()
+            >>> line_num = 0
+            >>> def try_parse_digit(c: str) -> Result[int, str]:
+            ...     try:
+            ...         res = Some(unicodedata.digit(c))
+            ...     except ValueError as e:
+            ...         res = NONE
+            ...     return res.ok_or(c)
+            >>>
+            >>> digit = iterator.next_if_map(try_parse_digit)
+            >>> while digit.is_some():
+            ...     line_num = line_num * 10 + digit.unwrap()
+            ...     digit = iterator.next_if_map(try_parse_digit)
+            >>> line_num
+            125
+            >>> iterator.join("")
+            ' GOTO 10'
+
+            ```
+        """
+        match self.next():
+            case Some(item):
+                match f(item):
+                    case Ok(result):
+                        return Some(result)
+                    case Err(item):
+                        unpeek = Some(item)
+            case Null():
+                unpeek = NONE
+
+        self._peeked = unpeek
+        return NONE